stripe-salesforce-int/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
stripe
requests

